se arma consulta de list puntos reciclaje

+++ src/main/java/org/circle8/controller/response/PuntoReciclajeResponse.java
@@ -10,12 +10,12 @@ import java.util.List;
 @AllArgsConstructor
 @Builder(toBuilder = true)
 public class PuntoReciclajeResponse implements ApiResponse {
-	public int id;
-	public float latitud;
-	public float longitud;
+	public long id;
+	public double latitud;
+	public double longitud;
 	public List<DiaResponse> dias;
 	public List<TipoResiduoResponse> tipoResiduo;
 	public String recicladorUri;
-	public Integer recicladorId;
+	public long recicladorId;
 	public CiudadanoResponse reciclador;
 }
